Cleanup and better organization of code
